Added a track temporarilty to GameManager and completed the AudioManager. Slight folder restructure. Still need to work out speed as that doesnt seem to be native to the library
