+++ Waves/Assets/Audio/AudioResources.cs
@@ -1,0 +1,26 @@
+namespace Waves.Assets.Audio;
+
+/// <summary>
+/// Provides type-safe access to embedded audio resources.
+/// </summary>
+public static class AudioResources
+{
+    /// <summary>
+    /// Background music tracks.
+    /// </summary>
+    public static class Music
+    {
+        /// <summary>
+        /// Main game background music track.
+        /// </summary>
+        public const string Waves_001 = "Assets/Audio/Music/Waves_001.wav";
+    }
+
+    /// <summary>
+    /// Sound effects.
+    /// </summary>
+    public static class SoundEffects
+    {
+        // Add sound effect constants here as needed
+    }
+}
