+++ Waves/Assets/EnemyAssets.cs
@@ -1,0 +1,16 @@
+using Waves.Assets.BaseAssets;
+
+namespace Waves.Assets;
+
+/// <summary>
+/// Enemy assets.
+/// </summary>
+public static class EnemyAssets
+{
+    public static readonly IAsset Basic = new SingleCharAsset('☺');
+
+    public static readonly IAsset Shark = new MultiCharAsset(
+        "  /|",
+        "~~~~~~"
+    );
+}
