--- Waves/Core/Assets/EnemyAssets.cs
@@ -1,16 +1,0 @@
-using Waves.Core.Assets.BaseAssets;
-
-namespace Waves.Core.Assets;
-
-/// <summary>
-/// Enemy assets.
-/// </summary>
-public static class EnemyAssets
-{
-    public static readonly IAsset Basic = new SingleCharAsset('☺');
-
-    public static readonly IAsset Shark = new MultiCharAsset(
-        "  /|",
-        "~~~~~~"
-    );
-}
