+++ Waves/Entities/BaseEntity.cs
@@ -1,4 +1,4 @@
-using Waves.Core.Assets.BaseAssets;
+using Waves.Assets.BaseAssets;
 using Waves.Core.Interfaces;
 using Waves.Core.Maths;
