+++ Waves/Program.cs
@@ -39,6 +39,7 @@ internal class Program
                 services.AddSingleton<ScoreSystem>(sp => new ScoreSystem(sp.GetRequiredService<IGameManager>()));
 
                 services.AddSingleton<IAudioManager, AudioManager>();
+                services.AddHostedService<AudioManager>();
 
                 // Register GameLoop with all game systems injected
                 services.AddSingleton<IGameLoop>(sp => new GameLoop(
